requirements.txt
flask
requests
readability-lxml
beautifulsoup4
